requirements.txt
# requirements.txt for FarmLink Backend

# Core Flask & extensions
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1

# Database & MySQL driver
SQLAlchemy==2.0.46
PyMySQL==1.1.2

# Serial communication (for sensors)
pyserial==3.5

# Production server (Render, Railway, etc.)
gunicorn==25.1.0

# Utilities
blinker==1.9.0
click==8.3.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.5

# Optional but recommended for environment variables (if not already using)
# python-dotenv==1.2.1   (already included)